drive practice update
As far as drivetrain features go: Added swerve turn optimization(untested but will work), essentially instead of the module doing a 180 when switching directions it will instead just spin the other way. This is done per module and not just for 180 degree rotations but pretty much for everything above 90 degrees away. Also renamed manualMove to move since it is also called during autonomous.